LPS-67544 Move Liferay-specific "compileJSP" configuration
