+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/JspCDefaultsPlugin.java
@@ -14,39 +14,27 @@
 
 package com.liferay.gradle.plugins;
 
-import com.liferay.gradle.plugins.extensions.LiferayExtension;
 import com.liferay.gradle.plugins.jasper.jspc.CompileJSPTask;
 import com.liferay.gradle.plugins.jasper.jspc.JspCPlugin;
-import com.liferay.gradle.plugins.tasks.WritePropertiesTask;
 import com.liferay.gradle.plugins.util.FileUtil;
 import com.liferay.gradle.plugins.util.GradleUtil;
-import com.liferay.gradle.util.Validator;
 
 import java.io.File;
-import java.io.IOException;
 
-import java.util.Properties;
 import java.util.concurrent.Callable;
 
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.gradle.api.UncheckedIOException;
 import org.gradle.api.file.ConfigurableFileCollection;
-import org.gradle.api.internal.plugins.osgi.OsgiHelper;
 import org.gradle.api.plugins.JavaPlugin;
-import org.gradle.api.tasks.TaskContainer;
 import org.gradle.api.tasks.bundling.Jar;
-import org.gradle.api.tasks.compile.JavaCompile;
 
 /**
  * @author Andrea Di Giorgi
  */
 public class JspCDefaultsPlugin
 	extends BasePortalToolDefaultsPlugin<JspCPlugin> {
-
-	public static final String JSP_PRECOMPILE_ENABLED_PROPERTY_NAME =
-		"jsp.precompile.enabled";
 
 	public static final String UNZIP_JAR_TASK_NAME = "unzipJar";
 
@@ -108,65 +96,9 @@ public class JspCDefaultsPlugin
 				@Override
 				public void execute(Project project) {
 					addDependenciesJspC(project);
-					configureTaskCompileJSP(project);
 				}
 
 			});
-	}
-
-	protected void configureTaskCompileJSP(Project project) {
-		boolean jspPrecompileEnabled = GradleUtil.getProperty(
-			project, JSP_PRECOMPILE_ENABLED_PROPERTY_NAME, false);
-
-		if (!jspPrecompileEnabled) {
-			return;
-		}
-
-		JavaCompile javaCompile = (JavaCompile)GradleUtil.getTask(
-			project, JspCPlugin.COMPILE_JSP_TASK_NAME);
-
-		String dirName = null;
-
-		TaskContainer taskContainer = project.getTasks();
-
-		WritePropertiesTask recordArtifactTask =
-			(WritePropertiesTask)taskContainer.findByName(
-				LiferayRelengPlugin.RECORD_ARTIFACT_TASK_NAME);
-
-		if (recordArtifactTask != null) {
-			Properties artifactProperties;
-
-			try {
-				artifactProperties = FileUtil.readProperties(
-					recordArtifactTask.getOutputFile());
-			}
-			catch (IOException ioe) {
-				throw new UncheckedIOException(ioe);
-			}
-
-			String artifactURL = artifactProperties.getProperty("artifact.url");
-
-			if (Validator.isNotNull(artifactURL)) {
-				int index = artifactURL.lastIndexOf('/');
-
-				dirName = artifactURL.substring(
-					index + 1, artifactURL.length() - 4);
-			}
-		}
-
-		if (Validator.isNull(dirName)) {
-			dirName =
-				_osgiHelper.getBundleSymbolicName(project) + "-" +
-					project.getVersion();
-		}
-
-		LiferayExtension liferayExtension = GradleUtil.getExtension(
-			project, LiferayExtension.class);
-
-		File dir = new File(
-			liferayExtension.getLiferayHome(), "work/" + dirName);
-
-		javaCompile.setDestinationDir(dir);
 	}
 
 	protected void configureTaskGenerateJSPJava(final Project project) {
@@ -213,7 +145,5 @@ public class JspCDefaultsPlugin
 	}
 
 	private static final String _PORTAL_TOOL_NAME = "com.liferay.jasper.jspc";
-
-	private static final OsgiHelper _osgiHelper = new OsgiHelper();
 
 }
